TaskPage
Стилизация по точкам перелома, попытка анимировать появление, скрытие модального окна.
