+++ src/components/TaskForm/TaskForm.css
@@ -3,18 +3,9 @@
   flex-direction: column;
 }
 
-.modalLabelTask {
-  padding-top: 50px;
-  padding-left: 50px;
-}
-
 .modalInputTask,
 .modalInputNumber {
-  width: 428px;
   font-family: var(--main-font);
-  font-size: 18px;
-  line-height: 0.82;
-  letter-spacing: 0.04em;
   color: rgba(24, 28, 39, 0.6);
   border: 1px solid rgba(24, 28, 39, 0.2);
   border-top-style: none;
@@ -23,13 +14,7 @@
 }
 
 .modalLabelNumber {
-  padding-top: 50px;
-  padding-left: 50px;
   margin-bottom: 40px;
-}
-
-.modalInputNumber {
-  width: 220px;
 }
 
 .modalButtonBlock {
@@ -42,9 +27,16 @@
   margin-bottom: 20px;
   padding: 11px 44px;
   background: var(--button-color);
-  /* width: 154px; */
   border: none;
   display: block;
+  font-family: var(--main-font);
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 1.22;
+  display: flex;
+  align-items: center;
+  text-align: center;
+  letter-spacing: 0.04em;
 
   margin-left: auto;
   margin-right: auto;
@@ -62,11 +54,67 @@
   cursor: pointer;
   font-family: var(--main-font);
   font-size: 18px;
-  line-height: 0.82;
+  line-height: 1.22;
   letter-spacing: 0.04em;
   border-top: 0px;
   border-right: 0px;
   border-bottom: 1px solid rgba(24, 28, 39, 0.6);
   border-left: 0px;
   background-color: inherit;
+}
+
+@media screen and (min-width: 321px) {
+  .modalInputTask {
+    width: 428px;
+  }
+  .modalInputNumber {
+    width: 220px;
+  }
+  .modalInputTask,
+  .modalInputNumber {
+    font-size: 18px;
+    line-height: 0.82;
+    letter-spacing: 0.04em;
+  }
+  .modalLabelTask {
+    padding-top: 50px;
+  }
+  .modalLabelNumber {
+    padding-top: 50px;
+  }
+}
+
+@media screen and (min-width: 769px) {
+  .modalForm {
+    padding-left: 50px;
+  }
+}
+
+@media (min-width: 321px) and (max-width: 768px) {
+  .modalForm {
+    padding-left: 32px;
+  }
+}
+
+@media screen and (max-width: 320px) {
+  .modalForm {
+    padding-left: 20px;
+  }
+  .modalInputTask {
+    width: 280px;
+  }
+  .modalInputTask,
+  .modalInputNumber {
+    font-size: 16px;
+    line-height: 1.25;
+  }
+  .modalInputNumber {
+    width: 280px;
+  }
+  .modalLabelTask {
+    padding-top: 30px;
+  }
+  .modalLabelNumber {
+    padding-top: 58px;
+  }
 }
